Added LEDs to indicate levels initialization. Removed smart pointers - can't get them working on Arduino Uno.

+++ include/SoundSample.h
@@ -9,7 +9,7 @@ const int notProcessed = -1;
 
 class SoundSample {
     int inputCount = 0;
-    int inputLevel = 0;
+    long inputLevel = 0;
 public:
     int processInput(int input);
